FIX: settup dropdowsn on staking
