+++ launcher/public/output.css
@@ -1,7 +1,7 @@
 @import url("https://fonts.googleapis.com/css2?family=Noto+Sans:wght@100;200;300;400;500;600;700;800;900&display=swap");
 
 /*
-! tailwindcss v3.4.4 | MIT License | https://tailwindcss.com
+! tailwindcss v3.4.3 | MIT License | https://tailwindcss.com
 */
 
 /*
@@ -1808,10 +1808,6 @@ video {
   height: 95%;
 }
 
-.h-\[95\.5\%\]{
-  height: 95.5%;
-}
-
 .h-fit{
   height: -webkit-fit-content;
   height: -moz-fit-content;
@@ -1824,6 +1820,10 @@ video {
 
 .h-screen{
   height: 100vh;
+}
+
+.h-\[95\.5\%\]{
+  height: 95.5%;
 }
 
 .max-h-10{
@@ -1976,6 +1976,10 @@ video {
 
 .max-h-full{
   max-height: 100%;
+}
+
+.max-h-\[409px\]{
+  max-height: 409px;
 }
 
 .min-h-11{
@@ -2318,6 +2322,14 @@ video {
   width: 100vw;
 }
 
+.w-64{
+  width: 16rem;
+}
+
+.w-\[230px\]{
+  width: 230px;
+}
+
 .min-w-\[100px\]{
   min-width: 100px;
 }
@@ -2370,6 +2382,12 @@ video {
 .flex-shrink-0{
   -ms-flex-negative: 0;
       flex-shrink: 0;
+}
+
+.flex-grow{
+  -webkit-box-flex: 1;
+      -ms-flex-positive: 1;
+          flex-grow: 1;
 }
 
 .origin-top-right{
@@ -2511,6 +2529,27 @@ video {
 .animate-spin{
   -webkit-animation: spin 1s linear infinite;
           animation: spin 1s linear infinite;
+}
+
+@-webkit-keyframes ping{
+  75%, 100%{
+    -webkit-transform: scale(2);
+            transform: scale(2);
+    opacity: 0;
+  }
+}
+
+@keyframes ping{
+  75%, 100%{
+    -webkit-transform: scale(2);
+            transform: scale(2);
+    opacity: 0;
+  }
+}
+
+.animate-ping{
+  -webkit-animation: ping 1s cubic-bezier(0, 0, 0.2, 1) infinite;
+          animation: ping 1s cubic-bezier(0, 0, 0.2, 1) infinite;
 }
 
 .cursor-default{
@@ -4841,6 +4880,10 @@ video {
   opacity: 0.8;
 }
 
+.opacity-75{
+  opacity: 0.75;
+}
+
 .shadow{
   --tw-shadow: 0 1px 3px 0 rgb(0 0 0 / 0.1), 0 1px 2px -1px rgb(0 0 0 / 0.1);
   --tw-shadow-colored: 0 1px 3px 0 var(--tw-shadow-color), 0 1px 2px -1px var(--tw-shadow-color);
@@ -5474,11 +5517,6 @@ html body {
   border-color: rgb(77 117 117 / var(--tw-border-opacity));
 }
 
-.hover\:border-amber-300:hover{
-  --tw-border-opacity: 1;
-  border-color: rgb(252 211 77 / var(--tw-border-opacity));
-}
-
 .hover\:border-blue-400:hover{
   --tw-border-opacity: 1;
   border-color: rgb(96 165 250 / var(--tw-border-opacity));
@@ -5527,6 +5565,11 @@ html body {
 .hover\:border-teal-600:hover{
   --tw-border-opacity: 1;
   border-color: rgb(13 148 136 / var(--tw-border-opacity));
+}
+
+.hover\:border-amber-300:hover{
+  --tw-border-opacity: 1;
+  border-color: rgb(252 211 77 / var(--tw-border-opacity));
 }
 
 .hover\:bg-\[\#212325\]:hover{
