+++ launcher/src/components/UI/edit-page/components/edit/setups/SetupDropdown.vue
@@ -74,7 +74,7 @@
     >
       <div
         v-if="isOpen"
-        class="absolute top-9 z-20 min-h-20 mt-1 origin-top-right rounded-sm shadow-md bg-gray-200 transition-all duration-100 divide-y-2 divide-gray-500 shadow-black"
+        class="absolute top-9 z-20 min-h-20 mt-1 origin-top-right shadow-md bg-gray-200 transition-all duration-100 divide-y divide-gray-400 shadow-black rounded-md p-1"
         :class="getDropdownWidth"
         @mouseleave="isOpen = false"
       >
@@ -84,7 +84,7 @@
           @click="selectServerView"
         >
           <img
-            class="col-start-1 col-span-1 w-5 h-5 rounded-full border border-gray-300 self-center bg-gray-100"
+            class="col-start-1 col-span-1 w-5 h-5 rounded-full border border-gray-200 self-center bg-gray-100"
             src="/img/icon/stereum-icons/stereum-logo.png"
             alt="Node Server View"
           />
@@ -94,10 +94,10 @@
           >
         </div>
         <div
-          v-for="setup in setupLists"
+          v-for="setup in setupsList"
           v-show="setup.setupId !== setupStore.selectedSetup?.setupId"
           :key="setup.setupName"
-          class="p-2 bg-gray-300 capitalize transition-colors duration-300 transform text-gray-600 hover:bg-blue-300 hover:text-gray-700 cursor-pointer text-sm font-bold overflow-hidden truncate grid grid-cols-6 gap-x-1"
+          class="p-2 bg-gray-200 capitalize transition-colors duration-300 transform text-gray-600 hover:bg-blue-300 hover:text-gray-700 cursor-pointer text-sm font-bold overflow-hidden truncate grid grid-cols-6 gap-x-1"
           :class="{
             'pointer-events-none opacity-50': noValidatorHandler(setup),
           }"
@@ -126,12 +126,14 @@
 </template>
 <script setup>
 import { useSetups } from "@/store/setups";
-import { computed, ref } from "vue";
+import { computed, ref, watch } from "vue";
 import { useRoute } from "vue-router";
 
 import RenameSetup from "./RenameSetup.vue";
+import { useServices } from "@/store/services";
+import { useDeepClone } from "@/composables/utils";
 
-const { list, newHeight } = defineProps({
+const { newHeight } = defineProps({
   list: {
     type: Array,
     required: true,
@@ -147,11 +149,31 @@ const emit = defineEmits(["selectRename", "selectSetup", "serverView", "confirmR
 
 const route = useRoute();
 const setupStore = useSetups();
+const serviceStore = useServices();
 
 const isOpen = ref(false);
 
-const setupLists = computed(() => {
-  return list.filter((s) => s.setupName !== "commonServices");
+const setupsList = computed(() => {
+  const validators = serviceStore.installedServices
+    .filter((s) => s.category === "validator")
+    .map((v) => v.service);
+
+  let output = setupStore.allSetups.filter((s) => s.setupName !== "commonServices");
+  output = output.map((setup) => {
+    if (!setup.services || setup.services.length === 0) {
+      setup.noValidator = true;
+    } else {
+      const hasValidator = setup.services.some((service) =>
+        validators.includes(service.service)
+      );
+      if (!hasValidator) {
+        setup.noValidator = true;
+      }
+    }
+
+    return setup;
+  });
+  return useDeepClone(output);
 });
 
 const getDropdownWidth = computed(() => {
@@ -180,6 +202,14 @@ const getSelectedOption = computed(() => {
 
   return option;
 });
+
+//watchers
+watch(
+  () => setupsList.value,
+  () => {
+    setupStore.stakingSetups = setupsList.value;
+  }
+);
 
 // Methods
